fix: separate base classes to separate files to support improved tree shaking across tools
