+++ packages/web-components/src/accordion-item/accordion-item.base.ts
@@ -1,0 +1,93 @@
+import { attr, FASTElement, nullableNumberConverter } from '@microsoft/fast-element';
+import { uniqueId } from '@microsoft/fast-web-utilities';
+import { toggleState } from '../utils/element-internals.js';
+
+/**
+ *
+ * @slot start - Content positioned before heading in the collapsed state
+ * @slot heading - Content which serves as the accordion item heading and text of the expand button
+ * @slot - The default slot for accordion item content
+ * @slot marker-expanded - The expanded icon
+ * @slot marker-collapsed - The collapsed icon
+ * @csspart heading - Wraps the button
+ * @csspart button - The button which serves to invoke the item
+ * @csspart content - The wrapper for the accordion item content
+ *
+ * @public
+ */
+export class BaseAccordionItem extends FASTElement {
+  /**
+   * The internal {@link https://developer.mozilla.org/docs/Web/API/ElementInternals | `ElementInternals`} instance for the component.
+   *
+   * @internal
+   */
+  public elementInternals: ElementInternals = this.attachInternals();
+
+  /**
+   * Configures the {@link https://www.w3.org/TR/wai-aria-1.1/#aria-level | level} of the
+   * heading element.
+   *
+   * @public
+   * @remarks
+   * HTML attribute: heading-level
+   */
+  @attr({
+    attribute: 'heading-level',
+    mode: 'fromView',
+    converter: nullableNumberConverter,
+  })
+  public headinglevel: 1 | 2 | 3 | 4 | 5 | 6 = 2;
+
+  /**
+   * Expands or collapses the item.
+   *
+   * @public
+   * @remarks
+   * HTML attribute: expanded
+   */
+  @attr({ mode: 'boolean' })
+  public expanded: boolean = false;
+
+  /**
+   * Handles expanded changes
+   * @param prev - previous value
+   * @param next - next value
+   */
+  public expandedChanged(prev: boolean, next: boolean): void {
+    toggleState(this.elementInternals, 'expanded', next);
+  }
+
+  /**
+   * Disables an accordion item
+   *
+   * @public
+   * @remarks
+   * HTML attribute: disabled
+   */
+  @attr({ mode: 'boolean' })
+  public disabled: boolean = false;
+
+  /**
+   * Handles disabled changes
+   * @param prev - previous value
+   * @param next - next value
+   */
+  public disabledChanged(prev: boolean, next: boolean): void {
+    toggleState(this.elementInternals, 'disabled', next);
+  }
+
+  /**
+   * The item ID
+   *
+   * @public
+   * @remarks
+   * HTML Attribute: id
+   */
+  @attr
+  public id: string = uniqueId('accordion-');
+
+  /**
+   * @internal
+   */
+  public expandbutton!: HTMLElement;
+}
