+++ packages/web-components/src/accordion-item/index.ts
@@ -1,4 +1,5 @@
-export { AccordionItem, BaseAccordionItem } from './accordion-item.js';
+export { BaseAccordionItem } from './accordion-item.base.js';
+export { AccordionItem } from './accordion-item.js';
 export type { AccordionItemOptions } from './accordion-item.js';
 export { AccordionItemSize, AccordionItemMarkerPosition } from './accordion-item.options.js';
 export { styles as accordionItemStyles } from './accordion-item.styles.js';
