+++ packages/web-components/src/anchor-button/index.ts
@@ -1,5 +1,6 @@
 export { definition as AnchorButtonDefinition } from './anchor-button.definition.js';
-export { BaseAnchor, AnchorButton } from './anchor-button.js';
+export { BaseAnchor } from './anchor-button.base.js';
+export { AnchorButton } from './anchor-button.js';
 export { AnchorButtonAppearance, AnchorButtonShape, AnchorButtonSize, AnchorTarget } from './anchor-button.options.js';
 export type { AnchorButtonOptions } from './anchor-button.options.js';
 export { template as AnchorButtonTemplate } from './anchor-button.template.js';
