+++ packages/web-components/src/button/index.ts
@@ -1,5 +1,6 @@
 export { definition as ButtonDefinition } from './button.definition.js';
-export { BaseButton, Button } from './button.js';
+export { BaseButton } from './button.base.js';
+export { Button } from './button.js';
 export { ButtonAppearance, ButtonFormTarget, ButtonShape, ButtonSize, ButtonType } from './button.options.js';
 export type { ButtonOptions } from './button.options.js';
 export { styles as ButtonStyles } from './button.styles.js';
