+++ packages/web-components/src/checkbox/index.ts
@@ -1,5 +1,6 @@
 export { definition as CheckboxDefinition } from './checkbox.definition.js';
-export { BaseCheckbox, Checkbox } from './checkbox.js';
+export { BaseCheckbox } from './checkbox.base.js';
+export { Checkbox } from './checkbox.js';
 export { CheckboxShape, CheckboxSize } from './checkbox.options.js';
 export type { CheckboxOptions } from './checkbox.options.js';
 export { styles as CheckboxStyles } from './checkbox.styles.js';
