+++ packages/web-components/src/field/index.ts
@@ -1,4 +1,5 @@
-export { BaseField, Field } from './field.js';
+export { BaseField } from './field.base.js';
+export { Field } from './field.js';
 export { LabelPosition as FieldLabelPosition, ValidationFlags } from './field.options.js';
 export type { SlottableInput } from './field.options.js';
 export { definition as FieldDefinition } from './field.definition.js';
