+++ packages/web-components/src/progress-bar/index.ts
@@ -1,5 +1,6 @@
 export { definition as ProgressBarDefinition } from './progress-bar.definition.js';
-export { BaseProgressBar, ProgressBar } from './progress-bar.js';
+export { BaseProgressBar } from './progress-bar.base.js';
+export { ProgressBar } from './progress-bar.js';
 export { ProgressBarShape, ProgressBarThickness, ProgressBarValidationState } from './progress-bar.options.js';
 export { styles as ProgressBarStyles } from './progress-bar.styles.js';
 export { template as ProgressBarTemplate } from './progress-bar.template.js';
