+++ packages/web-components/src/rating-display/index.ts
@@ -1,4 +1,5 @@
-export { BaseRatingDisplay, RatingDisplay } from './rating-display.js';
+export { BaseRatingDisplay } from './rating-display.base.js';
+export { RatingDisplay } from './rating-display.js';
 export { RatingDisplayColor, RatingDisplaySize } from './rating-display.options.js';
 export { template as RatingDisplayTemplate } from './rating-display.template.js';
 export { styles as RatingDisplayStyles } from './rating-display.styles.js';
