+++ packages/web-components/src/switch/switch.ts
@@ -1,5 +1,5 @@
 import type { StaticallyComposableHTML } from '../utils/index.js';
-import { BaseCheckbox } from '../checkbox/checkbox.js';
+import { BaseCheckbox } from '../checkbox/checkbox.base.js';
 
 export type SwitchOptions = {
   switch?: StaticallyComposableHTML<Switch>;
