+++ packages/web-components/src/tablist/index.ts
@@ -1,5 +1,6 @@
 export { definition as TablistDefinition } from './tablist.definition.js';
-export { BaseTablist, Tablist } from './tablist.js';
+export { BaseTablist } from './tablist.base.js';
+export { Tablist } from './tablist.js';
 export { TablistAppearance, TablistOrientation, TablistSize } from './tablist.options.js';
 export { styles as TablistStyles } from './tablist.styles.js';
 export { template as TablistTemplate } from './tablist.template.js';
