+++ packages/web-components/src/text-input/index.ts
@@ -1,5 +1,6 @@
 export { definition as TextInputDefinition } from './text-input.definition.js';
-export { BaseTextInput, TextInput } from './text-input.js';
+export { BaseTextInput } from './text-input.base.js';
+export { TextInput } from './text-input.js';
 export { TextInputAppearance, TextInputControlSize, TextInputType } from './text-input.options.js';
 export type { TextInputOptions } from './text-input.options.js';
 export { styles as TextInputStyles } from './text-input.styles.js';
